python_app/requirements.txt
PyQt6
opencv-python
pillow
numpy
pandas
